refactor(beta): split knowledge.py into knowledge/ package
One file per implementation: base, memory, disk, sqlite, redis, locked,
log, bootstrap, polling. Public API and __all__ unchanged.

Also fixes DiskKnowledgeStore and SqliteKnowledgeStore constructors to
accept str | os.PathLike[str] per AGENTS.md path rules.
